+++ src/components/Meal/Meals.js
@@ -1,7 +1,14 @@
+import { Fragment } from "react";
+import MealsList from "./MealsList";
 import MealsTitle from "./MealsTitle";
 
 const Meals = () => {
-	return <MealsTitle />;
+	return (
+		<Fragment>
+			<MealsTitle />
+			<MealsList />
+		</Fragment>
+	);
 };
 
 export default Meals;